% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available.mldrs}
\alias{available.mldrs}
\title{Obtain additional datasets available to download}
\usage{
available.mldrs()
}
\value{
A data.frame with the available multilabel datasets
}
\description{
\code{available.mldrs} retrieves the most up to date list of additional datasets. Those datasets are not
included into the package, but can be downloaded and saved locally.
}
\examples{
\dontrun{
library(mldr.datasets)
names <- available.mldrs()$Name
}
}
