% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{bibtex}
\alias{bibtex}
\title{Dataset with BibTeX entries}
\format{An mldr object with 7395 instances, 1836 attributes and 159 labels}
\source{
Katakis, I. and Tsoumakas, G. and Vlahavas, I., "Multilabel Text Classification for Automated Tag Suggestion", in Proc. ECML PKDD08 Discovery Challenge, Antwerp, Belgium, pp. 75-83, 2008
}
\usage{
bibtex()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
bibtex()  # Check and load the dataset
toBibtex(bibtex)
bibtex$measures
}
}
