% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\docType{data}
\name{birds}
\alias{birds}
\title{Dataset with sounds produced by birds and the species they belong to}
\format{An mldr object with 645 instances, 260 attributes and 19 labels}
\source{
Briggs, F. and Lakshminarayanan, B. and Neal, L. and Fern, X. Z. and Raich, R. and Hadley, S. J. K. and Hadley, A. S. and Betts, M. G., "Acoustic classification of multiple simultaneous bird species: A multi-instance multi-label approach", The Journal of the Acoustical Society of America, (6)131, pp. 4640--4650, 2012
}
\usage{
birds
}
\description{
Multilabel dataset from the sound domain.
}
\examples{
\dontrun{
toBibtex(birds)
birds$measures
}
}
\keyword{datasets}

