% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{bookmarks}
\alias{bookmarks}
\title{Dataset with data from web bookmarks and their categories}
\format{An mldr object with 87856 instances, 2150 attributes and 208 labels}
\source{
Katakis, I. and Tsoumakas, G. and Vlahavas, I., "Multilabel Text Classification for Automated Tag Suggestion", in Proc. ECML PKDD08 Discovery Challenge, Antwerp, Belgium, pp. 75-83, 2008
}
\usage{
bookmarks(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
bookmarks <- bookmarks()  # Check and load the dataset
toBibtex(bookmarks)
bookmarks$measures
}
}
