% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal500.R
\docType{data}
\name{cal500}
\alias{cal500}
\title{Dataset with music data along with labels for emotions, instruments, genres, etc.}
\format{An mldr object with 502 instances, 68 attributes and 174 labels}
\source{
Turnbull, Douglas and Barrington, Luke and Torres, David and Lanckriet, Gert, "Semantic annotation and retrieval of music and sound effects", Audio, Speech, and Language Processing, IEEE Transactions on, (2)16, pp. 467-476, 2008
}
\usage{
cal500
}
\description{
Multilabel dataset from the music domain.
}
\examples{
\dontrun{
toBibtex(cal500)
cal500$measures
}
}
\keyword{datasets}

