% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corel16k001.R
\name{corel16k001}
\alias{corel16k001}
\title{Datasets with data from the Corel image collection. There are 10 subsets in corel16k}
\format{An mldr object with 13766 instances, 500 attributes and 153 labels}
\source{
Barnard, K. and Duygulu, P. and Forsyth, D. and de Freitas, N. and Blei, D. M. and Jordan, M. I., "Matching words and pictures", Journal of Machine Learning Research, Vol. 3, pp. 1107--1135, 2003
}
\usage{
corel16k001()
}
\description{
Multilabel dataset from the image domain.
}
\examples{
toBibtex(corel16k001)
corel16k001$measures
}

