% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corel16k005.R
\name{corel16k005}
\alias{corel16k005}
\title{Datasets with data from the Corel image collection. There are 10 subsets in corel16k}
\format{An mldr object with 13847 instances, 500 attributes and 160 labels}
\source{
Barnard, K. and Duygulu, P. and Forsyth, D. and de Freitas, N. and Blei, D. M. and Jordan, M. I., "Matching words and pictures", Journal of Machine Learning Research, Vol. 3, pp. 1107--1135, 2003
}
\usage{
corel16k005()
}
\description{
Multilabel dataset from the image domain.
}
\examples{
\dontrun{
corel16k005()  # Check and load the dataset
toBibtex(corel16k005)
corel16k005$measures
}
}
