% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{corel16k006}
\alias{corel16k006}
\title{Datasets with data from the Corel image collection. There are 10 subsets in corel16k}
\format{An mldr object with 13859 instances, 500 attributes and 162 labels}
\source{
Barnard, K. and Duygulu, P. and Forsyth, D. and de Freitas, N. and Blei, D. M. and Jordan, M. I., "Matching words and pictures", Journal of Machine Learning Research, Vol. 3, pp. 1107--1135, 2003
}
\usage{
corel16k006(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the image domain.
}
\examples{
\dontrun{
corel16k006 <- corel16k006()  # Check and load the dataset
toBibtex(corel16k006)
corel16k006$measures
}
}
