% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{delicious}
\alias{delicious}
\title{Dataset generated from the del.icio.us site bookmarks}
\format{An mldr object with 16105 instances, 500 attributes and 983 labels}
\source{
Tsoumakas, G. and Katakis, I. and Vlahavas, I., "Effective and Efficient Multilabel Classification in Domains with Large Number of Labels", in Proc. ECML/PKDD Workshop on Mining Multidimensional Data, Antwerp, Belgium, MMD08, pp. 30--44, 2008
}
\usage{
delicious(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
delicious <- delicious()  # Check and load the dataset
toBibtex(delicious)
delicious$measures
}
}
