% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_measures.R
\name{density.level}
\alias{density.level}
\title{Calculate the density level of the dataset}
\usage{
\method{density}{level}(mld)
}
\arguments{
\item{mld}{An \code{"mldr"} object}
}
\description{
This function calculates the ratio of nonzero-valued elements over the total
of elements.
}
\examples{
density.level(emotions)
}
