% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{emotions}
\alias{emotions}
\title{Dataset with features extracted from music tracks and the emotions they produce}
\format{An mldr object with 593 instances, 72 attributes and 6 labels}
\source{
Wieczorkowska, A. and Synak, P. and Ra'{s}, Z., "Multi-Label Classification of Emotions in Music", Intelligent Information Processing and Web Mining, Vol. 35, Chap. 30, pp. 307-315, 2006
}
\usage{
emotions
}
\description{
Multilabel dataset from the music domain.
}
\examples{
\dontrun{
toBibtex(emotions)
emotions$measures
}
}
\keyword{datasets}
