% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurlexev_tra.R
\name{eurlexev_tra}
\alias{eurlexev_tra}
\title{List  with 10 folds of the train data from the EUR-Lex EUROVOC descriptors dataset}
\format{An mldr object with 17413 instances, 5000 attributes and 3993 labels}
\source{
Mencia, E. L. and Furnkranz, J., "Efficient pairwise multilabel classification for large-scale problems in the legal domain", Machine Learning and Knowledge Discovery in Databases, pp. 50--65, 2008
}
\usage{
eurlexev_tra()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
eurlexev_tra()  # Check and load the dataset
toBibtex(eurlexev_tra[[1]])
eurlexev_tra[[1]]$measures
}
}

