% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurlexsm_test.R
\name{eurlexsm_test}
\alias{eurlexsm_test}
\title{List  with 10 folds of the test data from the EUR-Lex subject matters dataset}
\format{An mldr object with 1935 instances, 5000 attributes and 201 labels}
\source{
Mencia, E. L. and Furnkranz, J., "Efficient pairwise multilabel classification for large-scale problems in the legal domain", Machine Learning and Knowledge Discovery in Databases, pp. 50--65, 2008
}
\usage{
eurlexsm_test()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
eurlexsm_test()  # Check and load the dataset
toBibtex(eurlexsm_test[[1]])
eurlexsm_test[[1]]$measures
}
}

