% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurlexsm_tra.R
\name{eurlexsm_tra}
\alias{eurlexsm_tra}
\title{List  with 10 folds of the train data from the EUR-Lex subject matters dataset}
\format{An mldr object with 17413 instances, 5000 attributes and 201 labels}
\source{
Mencia, E. L. and Furnkranz, J., "Efficient pairwise multilabel classification for large-scale problems in the legal domain", Machine Learning and Knowledge Discovery in Databases, pp. 50--65, 2008
}
\usage{
eurlexsm_tra()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
eurlexsm_tra()  # Check and load the dataset
toBibtex(eurlexsm_tra[[1]])
eurlexsm_tra[[1]]$measures
}

