% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\docType{data}
\name{flags}
\alias{flags}
\title{Dataset with features correspoinding to world flags}
\format{An mldr object with 194 instances, 19 attributes and 7 labels}
\source{
Goncalves, E. C. and Plastino, A. and Freitas, A. A., "A genetic algorithm for optimizing the label ordering in multi-label classifier chains", Tools with Artificial Intelligence (ICTAI), 2013 IEEE 25th International Conference on, pp. 469-476, 2013
}
\usage{
flags
}
\description{
Multilabel dataset from the image domain.
}
\examples{
\dontrun{
toBibtex(flags)
flags$measures
}
}
\keyword{datasets}

