% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genbase}
\alias{genbase}
\title{Dataset with genes data and their functional expression}
\format{An mldr object with 662 instances, 1186 attributes and 27 labels}
\source{
Diplaris, S. and Tsoumakas, G. and Mitkas, P. and Vlahavas, I., "Protein Classification with Multiple Algorithms", in Proc. 10th Panhellenic Conference on Informatics, Volos, Greece, PCI05, pp. 448--456, 2005
}
\usage{
genbase
}
\description{
Multilabel dataset from the biology domain.
}
\examples{
\dontrun{
toBibtex(genbase)
genbase$measures
}
}
\keyword{datasets}
