% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.mldr}
\alias{get.mldr}
\title{Get a multilabel dataset by name}
\usage{
get.mldr(name, download.dir = if (is.null(getOption("mldr.download.dir")))
  file.path(normalizePath("~"), ".mldr", "datasets") else
  getOption("mldr.download.dir"))
}
\arguments{
\item{name}{Name of the dataset to load}
}
\description{
\code{get.mldr} obtains a multilabel dataset, either by finding it in the download directory or downloading it.
}
\examples{
\dontrun{
library(mldr.datasets)
bibtex <- get.mldr("bibtex")
bibtex$measures
}
}
