% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{langlog}
\alias{langlog}
\title{Dataset with data from the Language forum discussion}
\format{An mldr object with 1460 instances, 1004 attributes and 75 labels}
\source{
Read, Jesse, "Scalable multi-label classification", University of Waikato, 2010
}
\usage{
langlog
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
toBibtex(langlog)
langlog$measures
}
}
\keyword{datasets}
