% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical.R
\docType{data}
\name{medical}
\alias{medical}
\title{Dataset generated from medical reports}
\format{An mldr object with 978 instances, 1449 attributes and 45 labels}
\source{
Crammer, K. and Dredze, M. and Ganchev, K. and Talukdar, P. P. and Carroll, S., "Automatic Code Assignment to Medical Text", in Proc. Workshop on Biological, Translational, and Clinical Language Processing,  Prague, Czech Republic, BioNLP07, pp. 129-136, 2007
}
\usage{
medical
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
toBibtex(medical)
medical$measures
}
}
\keyword{datasets}
