% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mldrs}
\alias{mldrs}
\title{Obtains and shows the list of additional datasets, those available to download}
\usage{
mldrs()
}
\description{
The function downloads from GitHub the most up to date list of additional datasets. Those are the datasets not
included into the package, but that can be downloaded and save locally.
}
\examples{

library(mldr.datasets)
mldrs()
}

