% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20ng.R
\docType{data}
\name{ng20}
\alias{ng20}
\title{Dataset with news messages and the news groups they belong to}
\format{An mldr object with 19300 instances, 1006 attributes and 20 labels}
\source{
Ken Lang, "Newsweeder: Learning to filter netnews", in Proc. 12th International Conference on Machine Learning, pp. 331-339, 1995
}
\usage{
ng20
}
\description{
Multilabel dataset from the text domain. The original name of the dataset is \code{20ng}
}
\examples{
\dontrun{
toBibtex(ng20)
ng20$measures
}
}
\keyword{datasets}
