% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfolds.R
\name{random.kfolds}
\alias{random.kfolds}
\title{Partition an mldr object into k folds}
\usage{
random.kfolds(mld, k = 5, seed = 10)
}
\arguments{
\item{mld}{The \code{mldr} object to be partitioned}

\item{k}{The number of folds to be generated. By default is 5}

\item{seed}{The seed to initialize the random number generator. By default is 10. Change it if you want to obtain partitions containing
different samples, for instance to use a 2x5 fcv strategy}
}
\value{
An \code{mldr.folds} object. This is a list containing k elements, one for each fold. Each element is made up
of two mldr objects, called \code{train} and \code{test}
}
\description{
This method randomly partitions the given dataset into k folds, providing training and test partitions for each fold.
}
\examples{
\dontrun{
library(mldr.datasets)
library(mldr)
folds.emotions <- random.kfolds(emotions)
summary(folds.emotions[[1]]$train)
summary(folds.emotions[[1]]$test)
}
}

