% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv1sub1.R
\name{rcv1sub1}
\alias{rcv1sub1}
\title{Dataset from the Reuters corpus (subset 1)}
\format{An mldr object with 6000 instances, 47236 attributes and 101 labels}
\source{
Lewis, D. D. and Yang, Y. and Rose, T. G. and Li, F., "RCV1: A new benchmark collection for text categorization research", The Journal of Machine Learning Research, Vol. 5, pp. 361-397, 2004
}
\usage{
rcv1sub1()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
toBibtex(rcv1sub1)
rcv1sub1$measures
}

