% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv1sub2.R
\name{rcv1sub2}
\alias{rcv1sub2}
\title{Dataset from the Reuters corpus (subset 2)}
\format{An mldr object with 6000 instances, 47236 attributes and 101 labels}
\source{
Lewis, D. D. and Yang, Y. and Rose, T. G. and Li, F., "RCV1: A new benchmark collection for text categorization research", The Journal of Machine Learning Research, Vol. 5, pp. 361-397, 2004
}
\usage{
rcv1sub2()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
toBibtex(rcv1sub2)
rcv1sub2$measures
}

