% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rcv1sub5}
\alias{rcv1sub5}
\title{Dataset from the Reuters corpus (subset 5)}
\format{An mldr object with 6000 instances, 47235 attributes and 101 labels}
\source{
Lewis, D. D. and Yang, Y. and Rose, T. G. and Li, F., "RCV1: A new benchmark collection for text categorization research", The Journal of Machine Learning Research, Vol. 5, pp. 361-397, 2004
}
\usage{
rcv1sub5(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
rcv1sub5 <- rcv1sub5()  # Check and load the dataset
toBibtex(rcv1sub5)
rcv1sub5$measures
}
}
