% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reutersk500.R
\name{reutersk500}
\alias{reutersk500}
\title{Dataset from the Reuters Corpus with the 500 most relevant features selected}
\format{An mldr object with 6000 instances, 500 attributes and 103 labels}
\source{
Read, Jesse, "Scalable multi-label classification", University of Waikato, 2010
}
\usage{
reutersk500()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
reutersk500()  # Check and load the dataset
toBibtex(reutersk500)
reutersk500$measures
}
}

