% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\docType{data}
\name{scene}
\alias{scene}
\title{Dataset from images with different natural scenes}
\format{An mldr object with 2407 instances, 294 attributes and 6 labels}
\source{
Boutell, M. and Luo, J. and Shen, X. and Brown, C., "Learning multi-label scene classification", Pattern Recognition, (9)37, pp. 1757--1771, 2004
}
\usage{
scene
}
\description{
Multilabel dataset from the image domain.
}
\examples{
\dontrun{
toBibtex(scene)
scene$measures
}
}
\keyword{datasets}

