% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slashdot.R
\docType{data}
\name{slashdot}
\alias{slashdot}
\title{Dataset generated from slashdot.org site entries}
\format{An mldr object with 3782 instances, 1079 attributes and 22 labels}
\source{
Read, J. and Pfahringer, B. and Holmes, G. and Frank, E., "Classifier chains for multi-label classification", Machine Learning, (3)85, pp. 333--359, 2011
}
\usage{
slashdot
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
toBibtex(slashdot)
slashdot$measures
}
}
\keyword{datasets}

