% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_measures.R
\name{sparsity.level}
\alias{sparsity.level}
\title{Calculate the sparsity level of the dataset}
\usage{
sparsity.level(mld)
}
\arguments{
\item{mld}{An \code{"mldr"} object}
}
\description{
This function calculates the ratio of zero-valued elements over the total of
elements. It is useful to decide whether to export in a dense or sparse format.
}
\examples{
sparsity.level(emotions)
}
