% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackex_chess.R
\docType{data}
\name{stackex_chess}
\alias{stackex_chess}
\title{Dataset from the Stack Exchange's chess forum}
\format{An mldr object with 1675 instances, 585 attributes and 227 labels}
\source{
Charte, Francisco and Rivera, Antonio J. and del Jesus, Maria J. and Herrera, Francisco, "QUINTA: A question tagging assistant to improve the answering ratio in electronic forums", in EUROCON 2015 - International Conference on Computer as a Tool (EUROCON), IEEE, pp. 1-6, 2015
}
\usage{
stackex_chess
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
toBibtex(stackex_chess)
stackex_chess$measures
}
}
\keyword{datasets}
