% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackex_coffee.R
\name{stackex_coffee}
\alias{stackex_coffee}
\title{Dataset from the Stack Exchange's coffee forum}
\format{An mldr object with 225 instances, 1763 attributes and 123 labels}
\source{
Charte, Francisco and Rivera, Antonio J. and del Jesus, Maria J. and Herrera, Francisco, "QUINTA: A question tagging assistant to improve the answering ratio in electronic forums", in EUROCON 2015 - International Conference on Computer as a Tool (EUROCON), IEEE, pp. 1-6, 2015
}
\usage{
stackex_coffee()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
stackex_coffee()
toBibtex(stackex_coffee)
stackex_coffee$measures
}
}
