% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackex_cooking.R
\name{stackex_cooking}
\alias{stackex_cooking}
\title{Dataset from the Stack Exchange's cooking forum}
\format{An mldr object with 10491 instances, 577 attributes and 400 labels}
\source{
Charte, Francisco and Rivera, Antonio J. and del Jesus, Maria J. and Herrera, Francisco, "QUINTA: A question tagging assistant to improve the answering ratio in electronic forums", in EUROCON 2015 - International Conference on Computer as a Tool (EUROCON), IEEE, pp. 1-6, 2015
}
\usage{
stackex_cooking()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
toBibtex(stackex_cooking)
stackex_cooking$measures
}

