% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackex_cs.R
\name{stackex_cs}
\alias{stackex_cs}
\title{Dataset from the Stack Exchange's computer science forum}
\format{An mldr object with 9270 instances, 635 attributes and 274 labels}
\source{
Charte, Francisco and Rivera, Antonio J. and del Jesus, Maria J. and Herrera, Francisco, "QUINTA: A question tagging assistant to improve the answering ratio in electronic forums", in EUROCON 2015 - International Conference on Computer as a Tool (EUROCON), IEEE, pp. 1-6, 2015
}
\usage{
stackex_cs()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
stackex_cs()  # Check and load the dataset
toBibtex(stackex_cs)
stackex_cs$measures
}
}
