% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmc2007.R
\name{tmc2007}
\alias{tmc2007}
\title{Dataset from airplanes failures reports}
\format{An mldr object with 28596 instances, 49060 attributes and 22 labels}
\source{
Srivastava, A. N. and Zane-Ulman, B., "Discovering recurring anomalies in text reports regarding complex space systems", Aerospace Conference, pp. 3853-3862, 2005
}
\usage{
tmc2007()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
tmc2007()  # Check and load the dataset
toBibtex(tmc2007)
tmc2007$measures
}

