% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_arts.R
\name{yahoo_arts}
\alias{yahoo_arts}
\title{Dataset generated from the Yahoo! web site index (arts category)}
\format{An mldr object with 7484 instances, 23146 attributes and 26 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_arts()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_arts()  # Check and load the dataset
toBibtex(yahoo_arts)
yahoo_arts$measures
}
}
