% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{yahoo_business}
\alias{yahoo_business}
\title{Dataset generated from the Yahoo! web site index (business category)}
\format{An mldr object with 11214 instances, 21924 attributes and 30 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_business(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_business <- yahoo_business()  # Check and load the dataset
toBibtex(yahoo_business)
yahoo_business$measures
}
}
