% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_computers.R
\name{yahoo_computers}
\alias{yahoo_computers}
\title{Dataset generated from the Yahoo! web site index (computers category)}
\format{An mldr object with 12444 instances, 34096 attributes and 33 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_computers()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_computers()  # Check and load the dataset
toBibtex(yahoo_computers)
yahoo_computers$measures
}
}
