% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_education.R
\name{yahoo_education}
\alias{yahoo_education}
\title{Dataset generated from the Yahoo! web site index (arts education)}
\format{An mldr object with 12030 instances, 27534 attributes and 33 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_education()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_education()  # Check and load the dataset
toBibtex(yahoo_education)
yahoo_education$measures
}
}

