% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_entertainment.R
\name{yahoo_entertainment}
\alias{yahoo_entertainment}
\title{Dataset generated from the Yahoo! web site index (arts entertainment)}
\format{An mldr object with 12730 instances, 32001 attributes and 21 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_entertainment()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
yahoo_entertainment()  # Check and load the dataset
toBibtex(yahoo_entertainment)
yahoo_entertainment$measures
}

