% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_health.R
\name{yahoo_health}
\alias{yahoo_health}
\title{Dataset generated from the Yahoo! web site index (health category)}
\format{An mldr object with 8205 instances, 30605 attributes and 32 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_health()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_health()  # Check and load the dataset
toBibtex(yahoo_health)
yahoo_health$measures
}
}

