% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{yahoo_recreation}
\alias{yahoo_recreation}
\title{Dataset generated from the Yahoo! web site index (recreation category)}
\format{An mldr object with 12828 instances, 30324 attributes and 22 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_recreation()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_recreation()  # Check and load the dataset
toBibtex(yahoo_recreation)
yahoo_recreation$measures
}
}
