% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_science.R
\name{yahoo_science}
\alias{yahoo_science}
\title{Dataset generated from the Yahoo! web site index (science category)}
\format{An mldr object with 6428 instances, 37187 attributes and 40 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_science()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_science()  # Check and load the dataset
toBibtex(yahoo_science)
yahoo_science$measures
}
}

