% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{yahoo_social}
\alias{yahoo_social}
\title{Dataset generated from the Yahoo! web site index (social category)}
\format{An mldr object with 12111 instances, 52350 attributes and 39 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_social(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_social <- yahoo_social()  # Check and load the dataset
toBibtex(yahoo_social)
yahoo_social$measures
}
}
