% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_society.R
\name{yahoo_society}
\alias{yahoo_society}
\title{Dataset generated from the Yahoo! web site index (society category)}
\format{An mldr object with 14512 instances, 31802 attributes and 27 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_society()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
yahoo_society()  # Check and load the dataset
toBibtex(yahoo_society)
yahoo_society$measures
}

