% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.R
\name{yeast}
\alias{yeast}
\title{Dataset with protein profiles and their categories}
\format{An mldr object with 2417 instances, 103 attributes and 14 labels}
\source{
Elisseeff, A. and Weston, J., "A Kernel Method for Multi-Labelled Classification", Advances in Neural Information Processing Systems, Vol. 14, pp. 681--687, 2001
}
\usage{
yeast()
}
\description{
Multilabel dataset from the biology domain.
}
\examples{
toBibtex(yeast)
yeast$measures
}

