% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_n_load.mldr}
\alias{check_n_load.mldr}
\title{Check if an mldr object is locally available and download it if needed}
\usage{
check_n_load.mldr(mldr.name)
}
\arguments{
\item{mldr.name}{Name of the dataset to load}
}
\description{
This function checks if the mldr object whose name is given as input is locally available, loading it in memory.
If necessary, the dataset will be downloaded from the GitHub repository and saved locally.
}
\examples{
\dontrun{
library(mldr.datasets)
check_n_load.mldr("bibtex")
bibtex$measures
}
}
