% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delicious.R
\name{delicious}
\alias{delicious}
\title{Dataset generated from the del.icio.us site bookmarks}
\format{An mldr object with 16105 instances, 500 attributes and 983 labels}
\source{
Tsoumakas, G. and Katakis, I. and Vlahavas, I., "Effective and Efficient Multilabel Classification in Domains with Large Number of Labels", in Proc. ECML/PKDD Workshop on Mining Multidimensional Data, Antwerp, Belgium, MMD08, pp. 30--44, 2008
}
\usage{
delicious()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
toBibtex(delicious)
delicious$measures
}

