% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{enron}
\alias{enron}
\title{Dataset with email messages and the folders where the users stored them}
\format{An mldr object with 1702 instances, 1001 attributes and 53 labels}
\source{
Klimt, B. and Yang, Y., "The Enron Corpus: A New Dataset for Email Classification Research", in Proc. ECML04, Pisa, Italy, pp. 217-226, 2004
}
\usage{
enron(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
enron <- enron()  # Check and load the dataset
toBibtex(enron)
enron$measures
}
}
