% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mldrs}
\alias{mldrs}
\title{Obtain and show a list of additional datasets available to download}
\usage{
mldrs()
}
\description{
The function downloads from GitHub the most up to date list of additional datasets. Those datasets are not
included into the package, but can be downloaded and saved locally.
}
\examples{
\dontrun{
library(mldr.datasets)
mldrs()
}
}

