% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{tmc2007_500}
\alias{tmc2007_500}
\title{Dataset from airplanes failures reports (500 most relevant features extracted)}
\format{An mldr object with 28596 instances, 500 attributes and 22 labels}
\source{
Srivastava, A. N. and Zane-Ulman, B., "Discovering recurring anomalies in text reports regarding complex space systems", Aerospace Conference, pp. 3853-3862, 2005
}
\usage{
tmc2007_500(...)
}
\arguments{
\item{...}{Additional options for the loading function (e.g. \code{download.dir})}
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
tmc2007_500 <- tmc2007_500()  # Check and load the dataset
toBibtex(tmc2007_500)
tmc2007_500$measures
}
}
