% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_reference.R
\name{yahoo_reference}
\alias{yahoo_reference}
\title{Dataset generated from the Yahoo! web site index (reference category)}
\format{An mldr object with 8027 instances, 39679 attributes and 33 labels}
\source{
Ueda, N. and Saito, K., "Parametric mixture models for multi-labeled text", Advances in neural information processing systems, pp. 721--728, 2002
}
\usage{
yahoo_reference()
}
\description{
Multilabel dataset from the text domain.
}
\examples{
\dontrun{
yahoo_reference()  # Check and load the dataset
toBibtex(yahoo_reference)
yahoo_reference$measures
}
}

